// packages/monorepo/src/nx-types.ts
export namespace Nx {
  /**
   * Configuration for nx targetDependencies.
   */
  export type TargetDependencies = { [target: string]: ITargetDependency[] };

  /**
   * Configuration for project specific targets.
   */
  export type ProjectTargets = { [target: string]: IProjectTarget };
// ...
  export type INamedInputs = { [name: string]: string[] };
// ...
  export interface ProjectConfig {
    /** Name of project */
    readonly name?: string;
    /** Project root dir path relative to workspace */
    readonly root?: string;

    /**
     * Named inputs
     * @see https://nx.dev/reference/nx-json#inputs-&-namedinputs
     */
    readonly namedInputs?: INamedInputs;

    /**
     * Targets configuration
     * @see https://nx.dev/reference/project-configuration
     */
    readonly targets?: ProjectTargets;

    /**
     * Project tag annotations
     *
     * @see https://nx.dev/reference/project-configuration#tags
     */
    readonly tags?: string[];

    /**
     * Implicit dependencies
     *
     * @see https://nx.dev/reference/project-configuration#implicitdependencies
     */
    readonly implicitDependencies?: string[];

    /**
     * Explicit list of scripts for Nx to include.
     * @see https://nx.dev/reference/project-configuration#ignoring-package.json-scripts
     */
    readonly includedScripts?: string[];
  }
// ...
}